fix(ember): move data getter of work item form to route
