+++ caluma/extensions/settings.py
@@ -47,4 +47,8 @@ settings.APPLICANT_TASK_SLUGS = [
     "additional-data-form",
 ]
 
+settings.ADDITIONAL_DATA_FORM = {
+    "periodika-antrag": "periodika-abrechnung",
+}
+
 settings.CASE_ID_CACHE_SECONDS = int(os.environ.get("CASE_ID_CACHE_SECONDS", "60"))
